@@ -8,10 +8,10 @@
 #ifndef ffmpeg_h
 #define ffmpeg_h
 
-int AVERR_EOF;
-int AVERR_DECODER_NOT_FOUND;
-int AVERR_STREAM_NOT_FOUND;
-int AVERR_EAGAIN;
-int AVERR_ENOMEM;
+int const AVERR_EOF;
+int const AVERR_DECODER_NOT_FOUND;
+int const AVERR_STREAM_NOT_FOUND;
+int const AVERR_EAGAIN;
+int const AVERR_ENOMEM;
 
 #endif /* ffmpeg_h */
